addresses most non-skg-related errors
residual-conflict-resolution:4

+++ solr/core/src/java/org/apache/solr/search/facet/CacheUpdateCountSlotAcc.java
@@ -143,7 +143,7 @@ final class CacheUpdateCountSlotAcc extends CountSlotAcc implements CacheUpdater
 
   @Override
   public Object getValue(int slotNum) throws IOException {
-    return topLevelCounts[slotNum];
+    return (long) topLevelCounts[slotNum];
   }
 
   @Override
